+++ src/tables/mod.rs
@@ -32,4 +32,5 @@ pub mod sbix;
 pub mod stat;
 pub mod vhea;
 pub mod vmtx;
+pub mod vorg;
 pub mod vvar;